introduce MessageFromUrl component
